myproject/requirements.txt
gunicorn
asgiref==3.8.1
bidict==0.23.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.3
Flask==3.0.2
Flask-SocketIO==5.3.6
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
pytest==8.1.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
simple-websocket==1.0.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.1
wsproto==1.2.0
